added more app settings to app.json. renamed some components due to bad naming. added three new image assets; splash loading screen asset, high res logo, icon logo.
